requirements.txt
django==4.0
djangorestframework==3.13.1
django-filter==21.1
drf-yasg==1.20.0
